This is adding basic server functions to view html
This is adding the base of the server by alowing it to present the html
file through the httpRequestHandler. Despite not needing ssl the firefox
browser was playing up and would not connect to a normal http server
without ssl. This means a ssl certificate has been and and implemented
into the server by it is not trusted by the browser so it gives a
certificate warning upon connecting.