Adds a way to send json file data to the server
This adds a post form to allow data upload of json files to the server
which should then store it in a database. Some basic instruction for
upload has been added as well user friendly. Theorectically this should
be able to handle multiple files but may need to add square brackets to
'data'.